requirements.txt
ado-py==0.0.7
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
lxml==4.6.4
multidict==5.2.0
psycopg-binary==3.0.3
psycopg2==2.9.2
pydantic==1.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
soupsieve==2.3.1
SQLAlchemy==1.3.15
typing==3.7.4.3
typing_extensions==4.0.0
urllib3==1.26.7
yarl==1.7.2